deploy/requirements.txt
tensorflow-serving-api==2.1.0